+++ hatch_pip_compile/installer.py
@@ -6,9 +6,6 @@ from abc import ABC, abstractmethod
 from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
-from hatchling.dep.core import dependencies_in_sync
-from packaging.requirements import Requirement
-
 if TYPE_CHECKING:
     from hatch_pip_compile.plugin import PipCompileEnvironment
 
@@ -36,18 +33,6 @@ class PluginInstaller(ABC):
         """
         self.install_dependencies()
 
-    def run_pip_compile(self) -> None:
-        """
-        Run pip-compile if necessary
-        """
-        if not self.environment.lockfile_up_to_date:
-            self.install_pip_tools()
-            if self.environment.piptools_lock_file.exists():
-                _ = self.environment.piptools_lock.compare_python_versions(
-                    verbose=self.environment.config.get("pip-compile-verbose", None)
-                )
-            self.environment.pip_compile_cli()
-
     def install_project(self) -> None:
         """
         Install the project (`--no-deps`)
@@ -70,21 +55,6 @@ class PluginInstaller(ABC):
                 )
             )
 
-    def install_pip_tools(self) -> None:
-        """
-        Install pip-tools (if not already installed)
-        """
-        with self.environment.safe_activation():
-            in_sync = dependencies_in_sync(
-                requirements=[Requirement("pip-tools")],
-                sys_path=self.environment.virtual_env.sys_path,
-                environment=self.environment.virtual_env.environment,
-            )
-            if not in_sync:
-                self.environment.platform.check_command(
-                    self.environment.construct_pip_install_command(["pip-tools"])
-                )
-
 
 class PipInstaller(PluginInstaller):
     """
@@ -96,7 +66,7 @@ class PipInstaller(PluginInstaller):
         Install the dependencies with `pip`
         """
         with self.environment.safe_activation():
-            self.run_pip_compile()
+            self.environment.run_pip_compile()
             if not self.environment.piptools_lock_file.exists():
                 return
             extra_args = self.environment.config.get("pip-compile-install-args", [])
@@ -118,7 +88,7 @@ class PipSyncInstaller(PluginInstaller):
         uninstall everything in the environment before deleting the
         lockfile.
         """
-        self.install_pip_tools()
+        self.environment.install_pip_tools()
         cmd = [
             self.environment.virtual_env.python_info.executable,
             "-m",
@@ -148,7 +118,7 @@ class PipSyncInstaller(PluginInstaller):
         3) (re)install project
         """
         with self.environment.safe_activation():
-            self.run_pip_compile()
+            self.environment.run_pip_compile()
             self.install_dependencies()
         if not self.environment.skip_install:
             if self.environment.dev_mode:
